Chad/fix imports (#34)
* Set up Heroku CD

* Cleanup and add master

* New lines and logging

* Change context of worker dockerfile

* Fix worker

* Clean up and remove docs stuff

* Move branch conditionality into travis yml

* Semicolon

* Switch ID to use eventbrites

* Handled duplicate events; added test for clobbers; added context manager for sessions; added database cleaner

* Used scoped_session

* Cleanup and newlines

* Remove init

* Move database config into models

* Added Geoalchemy for points

* Stashing

* Merge master

* Remove geoalchemy

* Remove geoalchemy
