+++ beluga/__init__.py
@@ -3,17 +3,17 @@ import logging
 
 from sanic import Sanic
 from sqlalchemy import create_engine
-from sqlalchemy.orm import scoped_session, sessionmaker
-from sqlalchemy.ext.declarative import declarative_base
 
 from beluga import config
 from beluga.routes import api
+from beluga.models import db_setup
+
 
 # Build app and configuration.
 app = Sanic()
 app.config.from_object(config)
 
-# Basic logging config.
+# Basic app level logging config.
 logging.basicConfig(level=logging.INFO)
 app.logger = logging.getLogger(__name__)
 
@@ -22,34 +22,6 @@ db_engine = create_engine(
     app.config.DATABASE_URL,
     convert_unicode=True)
 
-
-@contextmanager
-def session_scope():
-    """Provide a transactional scope around a series of operations."""
-    session = scoped_session(
-        sessionmaker(
-            autocommit=False,
-            autoflush=False,
-            bind=db_engine
-        )
-    )
-
-    try:
-        yield session
-        session.commit()
-    except:
-        session.rollback()
-        raise
-    finally:
-        session.close()
-
-
-# Database tables.
-with session_scope() as db_session:
-    Base = declarative_base()
-    Base.query = db_session.query_property()
-
-
 @app.middleware("request")
 async def log_uri(request):
     # Simple middleware to log the URI endpoint that was called
@@ -57,10 +29,8 @@ async def log_uri(request):
 
 
 @app.listener('before_server_start')
-async def before_server_start(app, loop):
-    app.logger.info("Standing tables")
-    import beluga.models  # noqa
-    Base.metadata.create_all(bind=db_engine)
+async def setup_db(app, loop):
+    db_setup()
 
 
 app.blueprint(api)
