+++ beluga/models.py
@@ -1,6 +1,52 @@
-import sqlalchemy as sa
+import logging
+import os
 
-from beluga import Base
+from contextlib import contextmanager
+
+import sqlalchemy as sa
+from sqlalchemy.ext.declarative import declarative_base
+
+
+# Base for all table.
+Base = declarative_base()
+
+# Database credentials.
+DATABASE_URL = os.environ.get('DATABASE_URL')
+
+# Basic model level logging config.
+logging.basicConfig(level=logging.INFO)
+logger = logging.getLogger(__name__)
+
+
+def get_engine():
+    """Creates a database engine from our database url string."""
+    return sa.create_engine(
+        DATABASE_URL,
+        convert_unicode=True)
+
+
+def db_setup():
+    """Stand tables."""
+    logger.info('Standing tables.')
+    Base.metadata.create_all(get_engine())
+
+
+@contextmanager
+def session_scope():
+    """Provide a transactional scope around a series of operations."""
+    session = sa.orm.scoped_session(
+        sa.orm.sessionmaker(
+            autocommit=False,
+            autoflush=False,
+            bind=get_engine()))
+    try:
+        yield session
+        session.commit()
+    except:
+        session.rollback()
+        raise
+    finally:
+        session.close()
 
 
 class User(Base):
